Flask_Project/requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-Cors
marshmallow
werkzeug
pycryptodome
